+++ extensions/easydict/src/releaseVersion/releaseNote.ts
@@ -2,47 +2,34 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-09-20 11:15
+ * @lastEditTime: 2022-09-25 16:54
  * @fileName: releaseNote.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const releaseNote = `
-## [v2.3.0] - 2022-09-20
-
-### ✨ 新功能
-
-- 添加 [Bing 翻译](https://www.bing.com/translator)。
-- 添加 Bing 语种识别。
-
-![Bing-1663603058](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/Bing-1663603058.png)
-
+## [v2.3.1] - 2022-09-25
 
 ### 💎 优化
 
-- 优化了启动速度，飞一般的感觉～
-- 优化了语种识别的速度和准确度。
+- 优化了图标显示。
 
 ### 🐞 修复
 
-- 修复了一些已知问题。
+- 修复了获取有道翻译 cookie 可能失败导致崩溃的问题。
+- 修复了 web 有道词典某些情况下不显示释义问题。
 
-#### 🍃 _如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧(^-^)_
+####  -  _如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧(^-^)_
 
 ---
 
-### ✨ New
-
-- Added support for [Bing Translator](https://www.bing.com/translator).
-- Added Bing language detection.
-
 ### 💎 Improvements
 
-- Improved startup speed, flying feeling~
-- Improved the speed and accuracy of language detection.
+- Improved icon display.
 
 ### 🐞 Fixes
 
-- Fixed some known issues.
+- Fixed a bug that caused extension to crash when getting the Youdao web cookie failed.
+- Fixed a bug that caused web Youdao dictionary to not display the definition.
 `;